requirements.txt
Flask
tensorflow
keras
pillow
numpy
opencv-python
